+++ src/main/java/xyz/kohara/stellarity/registry/item/CallOfTheVoid.java
@@ -21,6 +21,4 @@ public class CallOfTheVoid extends BowItem {
         super(properties);
     }
 
-
-    //TODO
 }